Remove node specific modules from SDK generator to be truly isomorphic (#20)
* Remove generation of unused request identifier

It seems that originally the team wanted to assign this identifier to
the x-flow-request-id header, but it was never completed. Yell at me
later if you want, but in my opinion this is silly and should not be a
concern of the SDK client for several reasons:

1. A correlation identifier is used for backend debugging. Shifting that
responsibility on the client is unfair.

2. This client, or any other like it, is not the only entry point to
the API (i.e. cUrl can be used to make calls to our API) therefore
shifting the responsibility on the client does not guarantee coverage.
We have proxy which in my opinion is more suitable for this task and the
x-flow-request-id header can be part of the response.

3. This is a UNIVERSAL client. The same code must run on the browser and
node without issues. However, we are pulling in modules such as Crypto
and Buffer which are not available on the browser. Ultimately, these
modules are either automatically or manually shimmed. We use Webpack
internally which adds nearly 1MB of source code to shim these modules.
This is obviously not great for web performance.

* Use btoa to base64 encode strings on browser environments

Buffer is a node specific feature and forces bundlers to shim for
browser environments, thereby bloating the JavaScript bundle downloaded by
browsers unnecessarily. Most modern browsers implement BTOA and when not
available users can polyfill the environment.