Make fetch function required (breaking)
Code is defaulting to an unbound fetch function that is expected to be
available in the global scope and throws an illegal invocation error.
Let the developer using the generated client choose which library to use
to perform requests. They can choose 'isomorphic-fetch', 'cross-fetch',
pass a bound fetch from the client side, etc.