Add ts_declarations_v2 generator
This generator actually generates declarations files, unlike the
"ts_declarations" generator, which generates a TypeScript module
exporting type information for the root service.

The intent is for users of this generator to include the generated
declaration files in their tsconfig.json file to include the type
information in their compilation.

Plan is to update existing TypeScript generators to demand the
generation of these declaration files. This will allow developers to use
the type declarations in their implementations without restrictions and
leverage the other TypeScript generators to enhance their development
experience. Also, generators that demand these declaration files won't
need to generate these types keeping them lean.

This is currently an experimental feature.